Adds Gulp task for handling styles
The `styles` gulp task provides an alternative to using pure CSS for
styling by introducing SASS to the project. All SASS files will compile
down to a single CSS file any time the `default` gulp task is run.

Included alongside SASS during compilation are a few gulp modules that
take care of basic front-end optimization techniques. These are
gulp-autoprefixer and gulp-cssnano. To better serve a stronger
development work flow, all compiled styles are renamed and added to a
separate folder location. This is a typical automated task technique
for working with SASS and it’s optimization counterparts. To aid this
process, gulp-rename was included as well.