Adds browser-sync support
Browser sync is a great tool for speeding up any kind of front-end
development pipeline. During development, any changes made within the
`src` folder (i.e. SASS or JS) will automagically update within the
browser. No need to refresh the browser manually any time a change is
made to a few simple styles. Script changes produce similar results.
Browser-sync offers a proxy option. While we are using Node to create a
server, Browser-sync will proxy against that Node instance and give way
to quicker edits on the front end. Running gulp `browser-sync` in and
of itself does not yield much. However, including it within the list of
tasks that execute when running `gulp` (i.e. the default task) allows
for automatic style and script updates on the fly. Browser-sync will
open a browser window automatically when the task is run and redirect
to a preconfigured port. The proxy is currently set to run from port
8777.

It’s important to note that now that browser-sync is running, the
actual Node server never has a chance to actually start. The next
branch will address this problem by introducing Nodemon to the
development workflow.