Adds nodemon support
Nodemon starts a Node Express server that was configured originally
with this repo. It monitors for changes, including app crashes and
other useful event logging. On change, nodemon will auto-reload the
server. If a failure keeps the server from restarting, nodemon will
provide a notification. Once the error is resolved, nodemon again
automatically reboots the Node server. With Nodemon, back-end
development remains continuos and speeds up dev time overall.

In conjunction with Browser-sync, the combination of both tools provide
a smooth dev workflow. Changes on the backend that create an event of
any kind, especially crashes, are logged and fixed with a server that
wants to keep trying to run. Changes on the front-end with browser-sync
acting as a proxy to the nodemon server gives the ability to develop on
the front-end, quickly, without worrying about what may or may not be
going on on the backend.

Refactoring is a must at this point before moving forward with feature
implementation. For now, this dev pipeline works and offers a little
something for everyone no matter if you’re more of a backend engineer
or front end engineer.